change muti-thread timer
